requirements.txt
flask==2.0.1
flask-cors==3.0.10
tweepy==4.12.1
python-dotenv==0.19.2
pymongo==4.3.3
langchain>=0.0.267
openai>=0.27.0
google-generativeai>=0.3.0
pinecone-client>=2.2.2
tavily-python>=0.2.6
uuid>=1.30
deepseek-ai>=0.3.0
